Extract StatePool out of migration precheck shim
A StatePool had been mixed with a State in the migration PrecheckShim
for convenience but this is a little confusing and makes future
changes harder.

This change also fixes the issue of Models being returned from
AllModels() which referred to State instances which had been
released. Instead, AllModelUUIDs() is now used in conjuction with a
StatePool.