propagate host's AZ to containers
pad.lv/1684325

Currently a provider-specific AZ is not propagated to containers
deployed on any provider - during container creation an empty
HardwareCharacteristics struct is initialized by a container manager.

It is important for certain applications to have this value exposed as
an environment variable in the hook context so that they can be
configured to be aware of physical placement with respect to failure
domains.

Given that provisioner has an ability to query machine state via
an apiserver facade, it is possible to get an availability zone for a
machine with a given tag. For that a new API call is introduced to query
that information.