storage: add "Provider.Releasable()" method
Add the Provider.Releasable() method, which
reports whether the storage provider supports
releasing storage. This is used in state to
prevent marking storage as "releasing", when
releasing that storage is not possible.

Currently, the Azure and Oracle storage
providers are the only two that do not
support releasing storage. Support for
Oracle should be introduced soon; support
for Azure is dependent on upstream changes.