provider/azure: fix volume detachment
When detaching the final volume from an Azure
machine, we were setting the VM's data disks field
to nil, rather than a pointer to an empty slice.
This was apparently ignored by Azure, so the data
disk remained attached. Setting the field so it
is a pointer to an empty slice causes the field
to be marshalled, and Azure detaches the disk.