provider/azure: use managed disks
This is a major change to the Azure provider,
to use managed disks. To use managed disks,
VMs must be placed in "aligned" availability
sets. Because of that, models support either
the old unmanaged disks or new managed disks,
but never both. Models created before this
change will continue to use unmanaged disks,
and new models will use managed disks.

The azure storage provider now has an optional
"account-type" attribute, which is the storage
account type: Standard_LRS or Premium_LRS.
The default is Standard_LRS. A defualt pool,
azure-premium, is added, with account-type set
to Premium_LRS.

For now, the model config's storage-account-type
only affects the storage type used for the root
disk of VMs in managed disk models. It continues
to affect all storage in unmanaged disk models.