Add-Model for CAAS Clouds
Adds the ability to create a new model in a CAAS cloud that has been
registered to an existing controller.

Creates a minimal model config, does not currently use existing
modelmanager modelconfigcreator, which has some IAAS-specific parts.

Adds CloudIsCAAS function to the cloud package to have one location to
determine if a cloud is a CAAS cloud type based on the type string
(which may be e.g. "kubernetes" or "dcos" etc)

Changes NewModel to not require CAAS model clouds to match the
controller's cloud, because the controller will be on an IAAS cloud in
the initial version of CAAS support. Eventually we may bootstrap a
controller onto the CAAS cloud itself.

Changes some TODOs which reference (unfortunately) former contributors.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>