state: introduce ModelOperation
Introduce state.ModelOperation, implementations of
which encapsulate the logic for building operations
for making changes to a model. Two implementations
are provided, for destroying units and applications.
Separating the building of operations from their
execution allows for customisation of the operations.
We could alternatively use parameter structs, as we
do elsewhere, but the approach taken should enable
us to work towards composable operations, as described
below.

For now, operations can only be executed in isolation,
via state.State.ApplyOperation. This is due to several
limitations:

 - It is unsafe to compose mgo/txn ops arbitrarily.
   We currently compose mgo/txn ops across collections,
   which is unsafe, but the changes are made in a way
   as to limit the damage of interrupted transactions,
   and these static compositions are tested.
 - There is an upper limit on how many operations can
   be safely included in a mgo/txn transaction, which
   we do sometimes reach.

Eventually, if/when we move to an alternative replication
mechanism such as Raft, then we can look at making these
operations composable. For example, it would make sense
to compose hypothetical AddApplication and Addunit
operations, making the "NumUnit" logic in AddApplication
redundant. This would enable us to move more of the
business logic out of the state package, keeping it
focused on model-checking.