destroy-{model,controller}: fix --destroy-storage
When destroying a pre-2.3 controller or model, we
require the user to specify --destroy-storage to
avoid users accidentally destroying storage. The
older controller will destroy storage without
prompting the user, so the client does the check.

This commit updates destroy-model and destroy-controller
to first check if there is any storage in the model
or controller, and only require the user to specify
--destroy-storage if there is storage. If there is
no storage, then continue to destroy the controller.