External controller updater worker
We introduce a new worker that watches for
addition and removal of external controller
entries, and then watches for changes to
each one's API addresses. When the addresses
change, we update the external controller
details in the local controller's database.

To support this worker, two new facades are
introduced: ExternalControllerUpdater, and
CrossController. The former is an internal
controller facade, used by the worker to
watch, fetch and store the local knowledge
of external controllers. The latter is an
anonymous controller facade that is used by
other controllers to watch and fetch the
controller's API connection info.

There's a drive-by fix to stop importing
apiserver/authentication in workers. Workers
should never be using backend code directly,
so we move the AnonymousUsername constant
to the api package.