Fix a bug which broke the non-official build path
We need to return the forced version from sync.buildAgentTarball,
otherwise when we merge the version info back into the available tools
we overwrite the build number (which needs to be used because there's no
version file in this case).

Unfortunately this breaks uploadSuite.TestUploadAndForceVersion and
uploadSuite.TestSyncAndForceVersion for reasons that I haven't been able
to understand.