Firewaller shouldn't fail if a machine isn't provisioned
When dealing with existing open ports, we should just skip machines that
would give a NotProvisionedError. It does appear that this only happens
in edge cases in production where things are failing to be cleaned up
(either lack of credentials, or something broken in a model), but having
the firewaller bounce all the time exacerbates the problem.

Related to bug #1722813