cert: drop bogus DNSName in Verify
When verifying certs, we were verifying
that the certificate would accept any
host name. There's no point in that,
as the client can just skip hostname
checking anyway, which would provide
the same (zero) value.

Drop the "anyServer" DNSName, which
fails due to a change in Go 1.9, where
the Common Name is ignored when there
are subject alternative names in the
certificate.