LXD Provider should be warning on unknown config
LXD Provider was not calling ValidateUnknownAttrs because it wasn't
actually providing any new attributes. However, this meant that it
failed to catch any typos in the config values for core config.

This makes it so that "juju bootstrap lxd --config typo=foo" will warn
that 'typo' is not a known field. Also because it is very common to make
a mistake wrt '_' vs '-' I added a check to lookup the name by an
obvious alternative spelling, and give the hint to the user.