state: destroying model removes application offers
Extend DestroyApplicationOperation with a flag
that allows the application to be destroyed
even when it has offers, removing the offers
at the same time as the application is removed.

The flag is set when running the application
removal cleanup that is scheduled when a model
is destroyed.

Fixes https://bugs.launchpad.net/juju/+bug/1724673