Also look for versions file in /usr/lib/juju
For snaps the versions file will be alongside the jujud binary in the
same directory, but when Juju is installed from a deb jujud will be in
/usr/bin, and we won't be able to install the version file there.
Instead, check /usr/lib/juju for the versions file if it's not beside
the binary. (If version files are in both locations the one alongside
the binary takes precedence.) The deb packaging will need to install the
versions file to that directory.