{api,apiserver}/singular: accept controller tags
Update the Singular API to accept controller tags
as well as model tags. Also rename the existing
"ControllerTag" to "ClaimantTag". The former predates
what we now know as controller tags, and really means
"machine tag of a controller agent". Renaming to
ClaimantTag gives us some room to allow for unit
agents if we want to extend the use of singular
any more.

The facade version has been bumped, but the old
version has been dropped entirely. Controllers
are always in sync with both their API client
and API server code, and this is a controller-only
API; there's no point in carrying the old version.