cmd/jujud/agent: controller-singular workers
Update the machine agent to support running
singular workers in the machine dependency
engine. We introduce two new flags:

 - is-controller-flag
 - is-responsible-controller-flag

The former is set for all controller machines;
the latter is set for at most one controller
machine at a time, by running a singular worker
claiming a lease for the controller.

The txnpruner, dblogpruner, and
externalcontrollerupdater workers are dependent
on the is-responsible-controller-flag input.