Require CAAS clouds to have one default region
This gives us a place in model defaults to store CAAS cloud specific
data on the controller that isn't controller-wide. (if you pass an empty
region to SetModelDefaults, it stores them with the ID "controller").

This hasn't been an issue for IAAS clouds because their cloud-specific
defaults are stored client-side in clouds.yaml, not in the DB.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>