rpc: pass context to methods
Methods may have an optional first parameter of a type
to which context.Context is assignable.

The RPC server has a context which is cancelled when
the connection is closed. We then create a child context
for each new request, which is cancelled when the request
completes.

In future we will want to extend the client so it can
accept a Context to Conn.Call, and send a message to the
server to cancel the request. This will also give us a
way of propagating OpenTracing data.