all: make lease/leadership calls cancellable
We currently rely on a hack to prevent blocking
API calls (waiting for lease/leadership to unblock)
from stopping the API server exiting cleanly. The
hack is to kill all of the workers run inside State.

Ideally we would move the workers outside of State,
but that's easier said than done. This gets us a
little closer to untangling the workers, since this
change will enable the API server to run as a manifold,
taking in the StateTracker's State, without too many
contortions.