state: upgrade leases before starting manager
Since the lease managers are started when you
open a State object, they may be in use before
the lease upgrade step runs. This can lead to
errors, since the lease manager will choke upon
seeing the old lease documents.

To work around this, we upgrade the lease docs
just prior to starting the lease manager. When
we extract the lease managers from State, we
can drop this bit of code and rely on the upgrade
step. The upgrade step has been left in place
for this purpose.

Fixes https://bugs.launchpad.net/juju/+bug/1728972