migration: Allow exporting a relation with no status
If a model has been migrated from a controller of a Juju version before
relation.Status was introduced, relations won't have status
records. Allow them to be migrated even if no status is found. We can't
easily add a status at import time because it's not clear what the
status should be.