state: fetch lease managers lazily
If a consumer obtains a lease/leadership
claimer from State, it should reasonably
expect that the claimer is valid until
the State object is closed. Currently,
the claimers are forever broken once the
internal lease manager worker exits.

We change the claimers returned by State
to lazily fetch the internal worker,
so that a call to Claim, etc. will get
a new internal worker if the old one
exited due to an error.