migrations: Check that units have relation scopes
We were seeing CI migration tests fail because the migration was being
started before subordinate units had entered their relation's scope. Add
a precheck for this situation so that the user can see that they need to
wait for the units to finish coming up (rather than the migration
aborting for a cryptic reason).