Cache being sequences that are 'old'.
It turns out that we actually have lots of ways to have old beings. We
don't want to have to read them from the database every time we do a
Sync. So instead we cache them as 'known-to-be-superseded', and we can
skip reading them from the database.

Note that this potentially leaves us with a gap in alive, because we
don't transition to the new seq for an entity until the previous one has
actually transitioned to dead, but I think that is true for the old code
as well.