Don't start migration until unit agents are idle
We were seeing failures because a migration would be started before
subordinate units had entered scope - this meant that the export would
fail. The migration source prechecks were updated to make the failure
occur earlier (https://github.com/juju/juju/pull/8018), but this test
also needs to be updated to wait for the unit agents to be finished,
otherwise it will fail because of the precheck.

An example:
http://qa.jujucharms.com/releases/5924/job/model-migration-lxd/attempt/2234