worker/state: change output to *StatePool
Change worker/state to output a *state.StatePool,
rather than *state.State, and make that worker
responsible for registering state metrics and
communicating the StatePool back up to the agent
for introspection.

The other workers are updated to take a StatePool.
Of particular note, worker/apiserver now accepts
a StatePool and passes that directly to the
apiserver.NewServer call. worker/apiserver no
longer contains a worker; its manifold just calls
through to apiserver.NewServer with the configuration
derived from manifold inputs.

The worker/state worker is now responsible for
pinging all of the state objects, and removing them
from the pool when the corresponding model becomes
Dead or is removed from state.