provider/maas: fix zone logic in node selection
Update the selectNode methods to accept a single
availability zone (or none), and return the special
ErrAvailabilityZoneFailed error iff a zone was
specified and acquiring a node in that zone failed
to find a node matching the constraints. This makes
the code simpler, and eliminates the code that
contained an off-by-one error.

Fixes https://bugs.launchpad.net/juju/+bug/1732021