Work in progress.
Start trying to pull all the stuff that is layered behind 20-different
abstractions to get all the way back to its-just-strings-in-a-database.
Instead of loading everything one-by-one and doing 100s of round trips
to the database, just pull it in a single query and update our data
structure.

Things are mostly progressing.