Quick improvement for bug #1732509.
Model.Users() does a lookup based on model-uuid, but we don't have an
index. So on controllers with many models (but maybe not many users in
any given model), it has to do a collection scan over all model*users
for every model, leading to model*(model*users) = M^2. This at least
improves that bit, though we can do much better with some rewrites.