storageprovisioner: ignore incoherent volumes
In Juju, volumes cannot be provisioned without
an attachment to a machine. This is because
volumes are typically required to exist in the
same availability zone as the machine; so we
create the volume in the same AZ as the machine
to which it is intitially attached.

Because of this, the storage provisioner and
storage providers assume that an unprovisioned,
Alive, volume will have an attachment. It has
been observed in the wild that this does not
always hold true.

This change prevents the storage provisioner
from panicking upon seeing such volumes, but
does nothing to prevent the incoherent state
from occurring in the first place. It is not
clear how this can happen yet; it is possibly
an artifact of upgrading from an older version.

Partially fixes https://bugs.launchpad.net/juju/+bug/1722818