Consider model in upgrade/restore login checks
When a controller machine agent is running
upgrade steps, or is restoring, we now check
the model when allowing through agents with
the same tag as the controller machine agent.

Move the upgrade/restore checks into the
apiserver package itself, and get rid of
the "LoginValidator". The agent just passes
in functions that the API server can call
to check the current upgrade or restore
status.

Removed some tests from the machine agent
related to restore status, which weren't really
testing anything meaningful. We should move the
restore watching code to the API server in 2.4,
and test it properly. See:
  https://bugs.launchpad.net/juju/+bug/1733552

Fixes https://bugs.launchpad.net/juju/+bug/1733250