state: destroy backing volumes with machines
When progressing a machine to Dead, where the
machine has volume-backed filesystems, ensure
we destroy the backing filesystem. The method
for removing machine-scoped filesystems had a
partial/broken approach to removing filesystems;
we now use the "removeFilesystemOps method for
consistency; a "removeVolumeOps" method has
been extracted, and is used similarly.