state: short-circuit remove machine filesystems
When a machine is destroyed and its cleanup is
run, short-circuit the removal of non-detachable
filesystems as long as the machine is non-manual.
For manual machines, the machine is not going
anywhere, so the filesystem must be detached and
destroyed by the machine's storage provisioner.

Note that the cleanup will still continue to fail
while the filesystem has a related a storage instance,
meaning the unit has not run the storage-detaching
hook. This ensures that we don't pull the filesystem
out from underneath the unit during destruction.

Fixes the other half of https://bugs.launchpad.net/juju/+bug/1722818