firewaller: improve firewaller egress watcher
The firewaller's egress address watcher has
been changed to wait for the initial event
of each watcher before doing sending any
output. It was previously waiting for some
initial events, but missed the relation units.
Due to this, tests would intermittently fail.

By waiting for the initial event of each
watcher, we can also eliminate the explicit
initialisation step in the watcher.