worker/caasoperator: download & unpack charm
When the operator starts up, it checks if it
has a charm directory. If it does not, it will
ask for the application's charm URL and SHA256
hash, download and verify, and then unpack it
into the charm directory. The agent also sets
its status to maintenance during that process,
and active later.