state: add CAASMode.(Set)ContainerSpec
Add methods for getting and setting the
container spec for applications and units
in CAAS models.

Fix a bug when creating applications in
CAAS models, causing the series to not
be inherited from the charm.

Also, add basic support for adding units
in a CAAS model. Very much incomplete,
just enough for testing the basics.