caas: add Broker.EnsureUnit method
This method accepts a unit name and
provider-specific "spec", and creates
or updates resources according to that
spec for the unit. The resources created
must be identifiable given the unit name,
so that we can later delete them.

Also, change the EnsureOperator signature
to just accept a *caas.OperatorConfig,
rather than a function to get one. It's
always going to be called, so the
indirection doesn't help.