Extracted the combined recorder from rpc.Conn
rpc.Conn now accepts a RecorderFactory rather than an
ObserverFactory. The Recorder interface has the same methods as
Observer, but they can return errors to stop the handling of the
request, which is needed for auditing.

I've kept the Observer interface, since the lack of errors makes
multiplexing simpler. The observer is now embedded in a combined
recorder that forwards messages to it but also passes them on to the
auditlog recorder, which has the opportunity to interrupt the request.