Fix feature test that broke because of auditlog
Previously the test was trying to log to /var/log/juju but continuing
even though it didn't have permissions (since it was only logging). Now
that audit logging will prevent logins unless it can record the login,
being unable to write to the logdir was causing the test to fail.

Also, ensure the test doesn't timeout if an assertion fails.