api/common: fix GetObservedNetworkConfig tests
GetObservedNetworkConfig was calling network.GetDefaultRoute
directly, so if the machine running the tests had a default
route on a device with the same name as one of those in the
test data, the results would be affected.

We update NetworkSource with a method for obtaining the
default route information, where the default implementation
just calls through to network.GetDefaultRoute. The tests
have a fake implementation.