rackspace: work around flaky cloud-init/apt update
When starting instances on RackSpace, cloud-init
frequently fails to resolve the APT archive hosts.
This leads to "apt-get update" failing, which later
leads to (for example) LXD containers not starting.

This appears to be a RackSpace networking issue.
We work around the problem by issuing our own
"apt-get update" as a bootcmd in cloud-init,
followed by another after a sleep if that fails.

Fixes https://bugs.launchpad.net/juju/+bug/1677425