state: use database name prefix
This means that we can potentially have two instances of State
using the same mongo instance, which is desired for testing
higher level functionality such as the JIMM service.

It also potentially means that we could have tests that
run in parallel or tests in different juju packages that
reuse the same MongoDB server instance rather than
starting a server for every package.

This is not everything that needs to be done - the SetAdminMongoPassword
method requires access to the admin database, which is
unique to a mongo server, so doing that is incompatible
with sharing. However most tests do not need to set the
admin password, so placing a hard requirement on non-shared
mongo because of that seems like overkill. Perhaps
in a future PR, the password-setting methods can be moved
out of the State type, so the State type always operates
within the DBPrefix constraints. For now, just make SetAdminMongoPassword
fail when there's a non-empty prefix set.