Add State.UpdateControllerConfig
This allows updating or unsetting a subset of controller config values
in state. At the moment this is limited to audit log settings - once the
work has been done to determine how other values are used this can be
broadened to more of them.

Removing a value results in the default value being used.