On Xenial, we don't actually want to fall back to juju-mongodb.
There was probably a transition period where this was useful, but we
actually don't want to be running juju-mongodb on Xenial.
See https://bugs.launchpad.net/juju/+bug/1743901 where we fell back to
juju-mongodb because there was a temporary problem with the archive
mirror. But eventually juju-mongodb3.2 was found and juju started trying
to use that instead of the old mongo 2.4, and that just breaks
everything because it tries to launch the new mongo on the old database.