worker/raft/rafttransport: introduce package
We introduce a new package, worker/raft/rafttransport,
which contains a manifold and worker which implements
raft.Transport. The transport operates over the Juju
API server.

The worker installs an HTTP handler into the API
server's mux at a configurable path, hijacking the
connections that request that endpoint. The handler
requires clients to authenticate as controllers.

The worker listens to the central hub for publication
of API server addresses, so it can map Juju machine
agent tags to addresses.

Parts of the code in this commit are based on
https://github.com/CanonicalLtd/raft-http.