apiserver: split HTTP server out of API server
This change does various (too many) things:

 - split the HTTP server (worker/httpserver) out of the API server
 - introduce Authenticator and Authorizer interfaces for
   authenticating/authorizing HTTP requests
 - split out a basic auth handler that wraps other
   handlers, and uses a provided Authenticator and Authorizer
 - introduces a table defining endpoints in apiserver/apiserver.go,
   along with which ones require authentication, authorizers to
   use, which to track, etc.
 - updates various apiserver tests to use a new base suite,
   rather than JujuConnSuite
 - drops various apiserver tests which repeat testing auth
   and model UUID validation, which are orthogonal of the
   endpoints being tested.

There are numerous benefits to this (IMO), but the primary goal
is to enable rafttransport to attach to the httpserver mux, allowing
the raft object to be fed into the apiserver worker.

Return authenticated entity in AuthInfo. The macaroon login tests were
failing because the finder.FindEntity calls would fail for remote
users. In that case the user doesn't exist in state, but the login
succeeds because we wrap a modelUserEntityFinder that handles them
correctly. Rather that requiring all of the users of the AuthInfo to
also use the modelUserEntityFinder to look up the tag, it seems better
to return the found entity directly (which is how the old authentication
function worked).

Unskip serverSuite.TestClosesStateFromPool. This test had suffered some
bit-rot - the API server no longer watches models and removes dead ones
from the pool (that's done by the state worker), so destroying a model
never removes it and the test always fails. Since the test now has
direct access to the state pool passed into the API server, it's much
simpler - it can issue the removal request itself.

Move agent download tests out to a feature test.

Move the tools macaroon tests to the tools feature test. This revealed a
bug in the discharge handling in the BasicAuthHandler - it needs to
return a json-serialised ErrorResult with details of the macaroon in the
response.

Fix the gui version tests - they were broken by overly-strict
assertResponse behaviour and a change in go 1.10 where it doesn't set
the content type automatically if there's no content. Allow the response
to have no content type if there's no response body. If there's a
content type set it still needs to match the expected value.

Handle macaroon discharge errors for streaming API clients: the
discharge error is serialised as params.ErrorResult and needs to be
unpacked from the response so that normal discharging can be done.

The mux now has AddClient and ClientDone methods that clients use to
indicate that they're using it (and when they've finished). The
httpserver won't shutdown until all clients have finished (at the moment
this is just the apiserver, but the rafttransport worker will be another
one).

Wait until handlers are added in apiserver.NewServer: We were seeing
intermittent failures with 404s when tests would try to connect to the
API, especially when they created multiple API connections - some would
fail while others would succeed. These were caused by the handlers not
having been added in loop yet.