Disallow changing audit-log-max-backups and max-size
Prevent these controller config settings from being changed after
bootstrap. Handling it correctly would require more work - one way to do
it would be to make a ReplaceableAuditLog that would wrap an
auditlog.LogFile and allow switching out the underlying one, closing it
and opening a new one. But that's not really needed at the moment and it
complicates things a fair bit.