Formatting fix for exclude method lists in controller-config
If there is more than one entry in the list, add a newline so all the
elements are shown below each other. If there's only one entry remove
the "- " from the start.

There's a pre-existing problem where newlines in the values mean
that the first column width shrinks back down after the value. This
change doesn't fix that.