Occasionally sockets get wedged in Mgo.
If they ever go into an abended state, they end up in the socket pool to
be used again, but even when the connection is reset, they don't exit
the abended state. This causes us to recheck isMaster on every query,
and that also causes us to reauthenticate on every query. (the pattern
in mgo is that every connection is queued for logout after every query,
but the next query adds a login and it collapses a logout + login pair
into a no-op. But isMaster breaks that logic and causes logout and login
on every query.)

This is a patch taken from:
    https://github.com/go-mgo/mgo/pull/255/files