Add manifold for auditconfigupdater worker
Pull cleanupWorker out into worker/common - leave the couple of
instances that need to be able to provide manifold outputs in place.

Expose auditconfigupdater outputs in manifold because state isn't
available in the machine agent manifolds until the state worker is
started - that means that we can't create the initial audit config
outside and pass it into the API server and config updater
independently. Instead create the config and the channel when the
manifold is started, and expose an output type that allows access to the
initial config and the change channel.

Update apiserver worker to use auditconfigupdater outputs.

Re-enable MachineSuite.TestManageModelAuditsAPI.