Handle the case where lease documents exist while migrating.
Under HA, it is possible to have the singleton lease document get created while
the upgrade steps are running. This would cause the migration code to fail,
because while the Create op was checking that the target didn't exist, we
weren't checking that the target didn't exist in the migration loop. Which
meant that we would try to delete the old and create the new, but fail the
assertion, and have no idea what was wrong.

Upon consideration, if we the lease we are 'migrating' already exists, then
someone has already taken the new lease, and it should win, and the old lease
just be discarded anyway. So that is what we do.