Rewrite the test so that it actually asserts based on real synchronization.
To force the issue, you can add this patch:
diff --git a/worker/periodicworker.go b/worker/periodicworker.go
index 27d0f94..c560bbc 100644
--- a/worker/periodicworker.go
+++ b/worker/periodicworker.go
@@ -108,6 +108,8 @@ func (w *periodicWorker) run(call PeriodicWorkerCall, period time.Duration) erro
                                return err
                        }
                }
+               // XXX: Force race conditions to become apparent, do *not* leave this in code
+               time.Sleep(time.Millisecond)
                timer.Reset(nextPeriod(period, w.jitter))
        }
 }

And see that the old test would fail, while the new test still passes.

The synchronization points are now:

a) We know that call() has been triggered 2 times because we waited for it.
b) We know that call() won't be called again because we've called Kill().
c) We know that all nextPeriod have been called because we Wait() for the
   Worker to notice the Kill() and exit, which means it has gone around its
   loop and called next period and gotten all the way around to <-stop.

I also confirmed that moving the w.Kill() after the last select properly fails
the test because it gets the wrong number of nextPeriods (it gets an extra one
late, and fails to record both of the earlier ones.)