Handle an interesting race condition (bug #1747367)
The static analysis showed that this might be a problem, and we're able to set
up a case where it really does fail.

Essentially, we have one loop which waits to try and send a notification of an
event to users. While waiting for that event to be sent, it allows new requests
to be made. So if we start watching for an event, but never handle when it
triggers, and then queue up a bunch more watch events that cause us to
reallocate the event slice, and *then* unwatch the original event, we will
never notice that the original event is no longer watched because we'll be hung
on an object that doesn't get updated.