Handle some small failures on Windows.
Mostly formatting issues that meant Windows improperly compared strings. One
test was using Matches when it wanted Equals, which didn't matter on Linux, but
on Windows the path section was interpreted as regex escapes. The other code
was improperly formatting code because of incorrect expectations about command
names. (On Windows they have a .exe extension.)