Properly sync the test TestOpenTimesOutOnLogin
It was using WaitAdvance, but we could actually be waiting in several
different locations, we were not guaranteed to be waiting in the Login
code. However, we already had a channel that we could use to know when
we got into the Login code, so use it in the test.