Avoid panic() when the test fails.
If checkStartInstance fails with an Assert failure, we weren't
forcing the goroutine to exit, which meant it could get a panic() after
we had already started tearing down the test, which would then suppress
the real error that was occurring.
This forces us to kill the goroutine before we return from the test.