Allow ForceDestroy to initiate teardown of Controllers.
We no longer prevent it from doing anything to a controller.
Instead our restriction is that you can only destroy a machine if
there are other controllers available. We cannot go to Dead if
we still HaveVote, but we can go to Dying, which should cause the
peergrouper to want to take us out of the voting set.