Simplify usage of apiserver/testserver
Previously clients needed to call StartTLS on the http server and
schedule closing of both the API and http servers. Wrap them up in a
server struct that stops both servers when asked.

(This was a review comment from @wallyworld while reviewing the
apiserver-httpserver split branch, but we decided it could go in a
smaller PR after the main one landed.)