Add a Tracef for all transactions.
Rather than tracing just the transactions that we rewrite, trace all
changes to the database. And do so using "pretty" which makes it easier
to understand the actual transactions.

We write it to a specific logger so that people can enable/disable just
the transaction logging.

Using pretty.Formatter should be less overhead.

Ideally we won't do any txn formatting if you don't have TRACE logging
enabled. Not sure if logger.Tracef does all the right things because it
may be enabled on one side but disabled in final writing.