Test suite now sometimes passes.
It seems the new test is not deterministic, so will need to be updated.
But it can pass, and does show that the peergrouper notices and cleans
up Dying controller machines after it has removed their vote.

Also implements a State.RemoveControllerMachine that allows you to run
the ops necessary to remove the jobs from a machine and remove it from
the list of machineids in the controller info.