Update to use the new txn.Runner that tracks Duration.
Since txn.Runner now makes Clock mandatory, we have to update all the
callers to pass in a clock.
This can't land as is, because gopkg.in/blobstore.v2 also needs the
updated juju/txn and pass in the Clock there as well.

But this can land as soon as we land the existing patch against
blobstore.