Make the test pass reliably.
workerSuite.TestDyingMachinesAreRemoved had a bit of a race around
noticing when the replicaset changed. We now have a fairly reliable set
of operations even though the worker loop is mostly independent of our
changes. (Especially that detecting Mongo changes is not driven by an
event that is easy to control from the test suite.)

But running:
  go test --test.failfast --test.count 100 --check.v --check.f Dying

Has it passing reliably.