Fix peergrouper internal address bug
The peergrouper was publishing server details with empty internal
addresses. This was because publishAPIServerDetails was expecting to be
passed a members map with all the machines' details in them, but
desiredPeerGroup would only return a non-nil members map if the set had
changed. updateReplicaSet would interpret the members map being nil to
indicate that the set hadn't changed.

This caused the raft workers to get all tangled up:

* the raft-transport's stream layer would wait for a non-blank
  address (streamLayer.Addr() blocks waiting for l.addr, but loop throws
  away the blank address so never writes it to l.addr).
* raft-worker would hang calling raft.NewRaft in loop() (because it
  calls streamLayer.Addr() for initialisation)
* which meant that raft-clusterer would hang in manifold.Start getting
  the raft dependency (since it calls raftWorker.Raft(), which needs
  loop to write to w.raftCh)

Change the desiredPeerGroup to return changed separately, and return
the members map even if there hasn't been a change.