Make peergrouper send current details on request
If the raft clusterer started after the peergrouper it would miss the
initial server details published on the hub, and never be able to
connect the raft nodes together. The peergrouper used to publish
apiserver details periodically to avoid this, but that still meant that
there could be quite a long wait before the details were republished.

Instead the peergrouper subscribes to apiserver.DetailsRequestTopic, and
once the raft-clusterer is subscribed to the details topic it publishes
a request. The peergrouper then publishes the current details in
response.

Also updated the pubsub forwarder to use this mechanism.