Prevent raft-transport address method hanging indefinitely
If a bug in the peergrouper meant that no useful server address is being
published, the raft transport's LocalAddr method would hang
forever (until controller shutdown) - this was difficult to track
down. Rather than that we timeout, returning a bad address to the raft
code and killing the transport worker (so the error should be very clear
in the log).