Move the FSM from rafttest to worker/raft
The FSM we're using for now was in the rafttest package - this imports
various test packages, and some setup they were doing at init time (I
haven't worked out what exactly) was causing problems in running the
uniter tests. For some reason it made the hooks commands run in the
tests much slower, and this meant that the tests would sometimes (about
50% of the time) hit timeouts and fail. It was a pain to track down
because the problem seemed to be caused by having the workers in the
machine agent manifold, but actually the tests are based on
JujuConnSuite and never run a dependency engine.

Moving the FSM to be worker/raft.SimpleFSM resolves fixes the tests.