add-apt-repository has been provided elsewhere for a long time.
As near as I can tell, since Trusty we should have been installing
software-properties-common instead of python-software-properties. We
just didn't notice because add-apt-repository is actually always
available (cloud-init depends on it in Trusty, ubuntu-server depends on
it in Xenial).

Bionic has removed the python-software-properties package last week, so
we shouldn't be trying to install it.