Move the LXD mock into our codebase.
The patch submitted by Witold was rejected in favor of using an
interface for Operations and not accepting the mock. But we can just as
easily create a local test mock which we use in our test infrastructure,
which mocks their API without it needing to be in their source tree.