Make sure enable-ha always wants 3/5/7 voters.
When we made it possible to remove a machine, that meant we could get a
'current' number of machines that want to vote to an even number (even
if the numebr of actual voters is kept to an odd number).

However, that meant that doing:
```
 $ juju enable-ha # gives you 3
 $ juju remove-machine 1
 # now have 1 voter (the primary), and another
 # that wants to vote
 $ juju enable-ha # should get you back to 3
```

However, with the old code the check would say "set the number to the
current number" which was 2. This changes it to always roll up to the
next odd number.

The other change in this patch are:

1. RemoveControllerMachine now properly makes sure it isn't removing the
   last controller. And we add 'race' tests that ensure we notice and give
   an error as to *why* we are unable to do what you asked (rather than
   just 'transaction aborted').

2. Machine.Destroy() also gets race tests that ensure we refuse to
   remove the last controller machine if another machine is also being
   removed.

3. While debugging with TEST_LOGGING_CONFIG="<root>=DEBUG;juju.state.txn=TRACE",
   I eventually noticed that transactions being run by the "BeforeHooks"
   functionality were not getting logged. This fixes that runner to also
   log to the txnLogger.
   Also, as an RFC, I put a summary of the transaction that is about to
   be run into juju.state.txn DEBUG. For these, I'm pretty sure that
   logging level is ok, because these are not frequent transactions.
   I'd like to give more 'rationale' for transactions, but that probably
   needs to be a different rework.